Jisho Yoake and Remove Dark Theme CSS
Wekateka updated CSS to apply to mobile devices. I have taken his fix.
When removing the dark theme, dark to light, in the iframe, the logic
to do so was bad. We now use indexOf instead of the last link tag in
content_script.js. After reloading the extension, the injected link tag
could be before the last link tag added naturally by the site page.